fmt: Move reload test to a more proper place
