+++ tokio-trace-fmt/src/filter/reload.rs
@@ -84,3 +84,56 @@ impl fmt::Display for Error {
 }
 
 impl error::Error for Error {}
+
+
+#[cfg(test)]
+mod test {
+    use ::*;
+    use std::ync::atomic::{AtomicUsize, Ordering},
+    use tokio_trace_core::{dispatcher::{self, Dispatch}, Metadata};
+
+    #[test]
+    fn reload_handle() {
+        static FILTER1_CALLS: AtomicUsize = AtomicUsize::new(0);
+        static FILTER2_CALLS: AtomicUsize = AtomicUsize::new(0);
+
+        fn filter1<N>(_: &Metadata, _: &span::Context<N>) -> bool {
+            FILTER1_CALLS.fetch_add(1, Ordering::Relaxed);
+            true
+        }
+
+        fn filter2<N>(_: &Metadata, _: &span::Context<N>) -> bool {
+            FILTER2_CALLS.fetch_add(1, Ordering::Relaxed);
+            true
+        }
+
+        fn event() {
+            trace!("my event");
+        }
+
+        let subscriber = FmtSubscriber::builder()
+            .with_filter(filter1 as fn(&Metadata, &span::Context<_>) -> bool)
+            .with_filter_reloading();
+        let handle = subscriber.reload_handle();
+        let subscriber = Dispatch::new(subscriber.finish());
+
+        dispatcher::with_default(&subscriber, || {
+            assert_eq!(FILTER1_CALLS.load(Ordering::Relaxed), 0);
+            assert_eq!(FILTER2_CALLS.load(Ordering::Relaxed), 0);
+
+            event();
+
+            assert_eq!(FILTER1_CALLS.load(Ordering::Relaxed), 1);
+            assert_eq!(FILTER2_CALLS.load(Ordering::Relaxed), 0);
+
+            handle
+                .reload(filter2 as fn(&Metadata, &span::Context<_>) -> bool)
+                .expect("should reload");
+
+            event();
+
+            assert_eq!(FILTER1_CALLS.load(Ordering::Relaxed), 1);
+            assert_eq!(FILTER2_CALLS.load(Ordering::Relaxed), 1);
+        })
+    }
+}
